Rename project folder "classic_algorithms_suite" to "ClassicAlgorithmsSuite" for improved consistency
