+++ src/main/java/com/github/netmastermichael/ClassicAlgorithmsSuite/SortingAlgorithms/SortingAlgorithmMetrics.java
@@ -1,4 +1,4 @@
-package com.github.netmastermichael.classic_algorithms_suite.SortingAlgorithms;
+package com.github.netmastermichael.ClassicAlgorithmsSuite.SortingAlgorithms;
 
 public class SortingAlgorithmMetrics {
